requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
frozenlist==1.7.0
idna==3.10
multidict==6.4.4
propcache==0.3.2
python-dotenv==1.1.0
soupsieve==2.7
typing_extensions==4.14.0
yarl==1.20.1
